Fix proteksi halaman tambah data menggunakan cookies

+++ api/hapus_pemasukan.php
@@ -1,8 +1,14 @@
 <?php
+if (!isset($_COOKIE['login']) || $_COOKIE['login'] !== "true") {
+    header("Location: /index.php");
+    exit;
+}
+
 include "koneksi.php";
 
 $id = $_GET['id'];
-
 mysqli_query($conn, "DELETE FROM pemasukan WHERE id='$id'");
 
 header("Location: pemasukan.php");
+exit;
+?>
